Implement MCP server routes, feature flags, and related tests
