+++ forge/ee/lib/index.js
@@ -52,6 +52,9 @@ module.exports = fp(async function (app, opts) {
         // Set the expert assistant Feature Flag
         app.config.features.register('expertAssistant', isAiEnabled && (app.config?.expert?.enabled ?? false), true)
 
+        // Set the expert platform automation Feature Flag (MCP platform tools server)
+        app.config.features.register('expertPlatformAutomation', isAiEnabled && (app.config?.expert?.enabled ?? false), true)
+
         // temporary until FF Expert Insights can be enabled on Self Hosted EE instance
         const isInsightsEnabled = isAiEnabled && app.config?.expert?.enabled && app.config?.expert?.insights?.enabled
         app.config.features.register('expertInsights', isInsightsEnabled ?? false, false)
