+++ package.json
@@ -76,6 +76,7 @@
         "@heroicons/vue": "2.1.5",
         "@levminer/speakeasy": "^1.4.2",
         "@node-red/util": "^5.0.0",
+        "@modelcontextprotocol/sdk": "^1.29.0",
         "@node-saml/passport-saml": "^5.0.0",
         "@redis/client": "^6.0.0",
         "@sentry/node": "^10.54.0",
@@ -125,6 +126,7 @@
         "vue-shepherd": "^3.0.0",
         "vue3-google-login": "^2.0.33",
         "yaml": "^2.3.1",
+        "zod": "^4.4.3",
         "zxcvbn": "^4.4.2"
     },
     "devDependencies": {
